+++ src/pages/mypage/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as MyInfo } from './myinfo';
+export { default as UserEdit } from './user-edit';
